Имените названия функций
