+++ js/constants.js
@@ -60,11 +60,6 @@ const ITEM_TYPES = {
   hotel: 'Отель',
 };
 
-const MAP_CANVAS = document.querySelector('#map-canvas');
-const SIMILAR_CARD_TEMPLATE = document.querySelector('#card')
-  .content
-  .querySelector('.popup');
-
 export {
   MIN_INDEX,
   NUMBER_MIN,
@@ -81,7 +76,5 @@ export {
   FEATURES,
   PHOTOS,
   ITEM_TYPES,
-  Location,
-  MAP_CANVAS,
-  SIMILAR_CARD_TEMPLATE
+  Location
 };
