+++ js/popup.js
@@ -1,16 +1,22 @@
 import {
-  ITEM_TYPES,
-  MAP_CANVAS,
-  SIMILAR_CARD_TEMPLATE
+  ITEM_TYPES
 } from './constants.js';
 
 import {
-  isPresenceElement,
-  addUrlPhoto,
   creatPluralNames
 } from './util.js';
 
-const renderAds = (data) => {
+import {
+  removeExtra,
+  fillPhotoOrDelete
+} from './dom-util.js';
+
+const MAP_CANVAS = document.querySelector('#map-canvas');
+const SIMILAR_CARD_TEMPLATE = document.querySelector('#card')
+  .content
+  .querySelector('.popup');
+
+const renderAd = (data) => {
   const cardElement = SIMILAR_CARD_TEMPLATE.cloneNode(true);
 
   cardElement.querySelector('.popup__type').textContent = ITEM_TYPES[data.offer.type];
@@ -25,17 +31,17 @@ const renderAds = (data) => {
   const modifiers = data.offer.features.map((features) => `popup__feature--${features}`);
   const popupFeature = cardElement.querySelectorAll('.popup__feature');
 
-  isPresenceElement(popupFeature, modifiers);
+  removeExtra(popupFeature, modifiers);
 
   cardElement.querySelector('.popup__description').textContent = data.offer.description;
 
   const POPUP_PHOTOS = cardElement.querySelector('.popup__photos');
   const POPUP_PHOTO = POPUP_PHOTOS.querySelector('.popup__photo');
 
-  addUrlPhoto(data, POPUP_PHOTOS, POPUP_PHOTO);
+  fillPhotoOrDelete(data.offer.photos, POPUP_PHOTOS, POPUP_PHOTO);
   MAP_CANVAS.appendChild(cardElement);
 };
 
 export {
-  renderAds
+  renderAd
 };
